+++ coupon_multiplier_free_product/README.rst
@@ -1,6 +1,6 @@
-===================================
-Sale Coupon Multiplier Free Product
-===================================
+==============================
+Coupon Multiplier Free Product
+==============================
 
 .. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
    !! This file is generated by oca-gen-addon-readme !!
@@ -14,10 +14,10 @@ Sale Coupon Multiplier Free Product
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsale--promotion-lightgray.png?logo=github
-    :target: https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_multiplier_free_product
+    :target: https://github.com/OCA/sale-promotion/tree/15.0/coupon_multiplier_free_product
     :alt: OCA/sale-promotion
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-sale_coupon_multiplier_free_product
+    :target: https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-coupon_multiplier_free_product
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
     :target: https://runboat.odoo-community.org/webui/builds.html?repo=OCA/sale-promotion&target_branch=15.0
@@ -25,7 +25,7 @@ Sale Coupon Multiplier Free Product
 
 |badge1| |badge2| |badge3| |badge4| |badge5| 
 
-This module extends from *coupon_multiplier_free_product* and allows to make
+Base module that sets the configurations to create
 promotions of the type 3x2 or take this product and get
 this one another that apply as many times as the promotion rules fulfill.
 
@@ -34,11 +34,31 @@ this one another that apply as many times as the promotion rules fulfill.
 .. contents::
    :local:
 
-Usage
-=====
+Configuration
+=============
 
-Follow the configurations from the base module and be sure to guarantee the
-promotions defined conditions.
+If you want to create a promotion program of the type 3x2:
+
+  #. Go to *Sales > Products > Promotion programs* and create a new one.
+  #. Set the reward type to *Multiple of*.
+  #. Set the *Free product* to the one in the promotion.
+  #. Set the *Force rewarded product* option.
+  #. Set the *Quantity* to reward.
+  #. Set the *Quantity* to fulfill the reward condition.
+  #. If you want a limit on how many times the product should be rewarded, set a
+     *Maximum reward quantity*.
+
+If you want to create a promotion program of the type take one of these products and
+get this one another for free:
+
+  #. Go to *Sales > Products > Promotion programs* and create a new one.
+  #. Set the reward type to *Multiple of*.
+  #. Set the *Free product* to the one in the promotion.
+  #. Set the *Quantity* to reward.
+  #. Set the *Quantity* to fulfill the reward condition.
+  #. Set the products domain to fulfill the reward condition.
+  #. If you want a limit on how many times the product should be rewarded, set a
+     *Maximum reward quantity*.
 
 Bug Tracker
 ===========
@@ -46,7 +66,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/sale-promotion/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/sale-promotion/issues/new?body=module:%20sale_coupon_multiplier_free_product%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/sale-promotion/issues/new?body=module:%20coupon_multiplier_free_product%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -87,6 +107,6 @@ Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
 
 |maintainer-chienandalu| 
 
-This module is part of the `OCA/sale-promotion <https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_multiplier_free_product>`_ project on GitHub.
+This module is part of the `OCA/sale-promotion <https://github.com/OCA/sale-promotion/tree/15.0/coupon_multiplier_free_product>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
