+++ coupon_multiplier_free_product/__init__.py
@@ -1,0 +1,1 @@
+from . import models
