+++ coupon_multiplier_free_product/__manifest__.py
@@ -1,7 +1,7 @@
 # Copyright 2021 Tecnativa - David Vidal
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 {
-    "name": "Sale Coupon Multiplier Free Product",
+    "name": "Coupon Multiplier Free Product",
     "summary": "Allows to configure a coupon for e-commerce only",
     "version": "15.0.1.0.0",
     "development_status": "Beta",
@@ -10,5 +10,6 @@
     "author": "Tecnativa, Odoo Community Association (OCA)",
     "maintainers": ["chienandalu"],
     "license": "AGPL-3",
-    "depends": ["coupon_multiplier_free_product"],
+    "depends": ["sale_coupon_order_line_link"],
+    "data": ["views/coupon_program_views.xml"],
 }
