+++ coupon_multiplier_free_product/models/__init__.py
@@ -1,0 +1,3 @@
+from . import coupon
+from . import coupon_program
+from . import coupon_reward
