+++ coupon_multiplier_free_product/readme/CONTRIBUTORS.rst
@@ -1,0 +1,4 @@
+* `Tecnativa <https://www.tecnativa.com>`_:
+
+  * David Vidal
+  * Stefan Ungureanu
