+++ coupon_multiplier_free_product/readme/DESCRIPTION.rst
@@ -1,3 +1,3 @@
-This module extends from *coupon_multiplier_free_product* and allows to make
+Base module that sets the configurations to create
 promotions of the type 3x2 or take this product and get
 this one another that apply as many times as the promotion rules fulfill.
