+++ coupon_multiplier_free_product/static/description/index.html
@@ -4,7 +4,7 @@
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
 <meta name="generator" content="Docutils: http://docutils.sourceforge.net/" />
-<title>Sale Coupon Multiplier Free Product</title>
+<title>Coupon Multiplier Free Product</title>
 <style type="text/css">
 
 /*
@@ -360,21 +360,21 @@ ul.auto-toc {
 </style>
 </head>
 <body>
-<div class="document" id="sale-coupon-multiplier-free-product">
-<h1 class="title">Sale Coupon Multiplier Free Product</h1>
+<div class="document" id="coupon-multiplier-free-product">
+<h1 class="title">Coupon Multiplier Free Product</h1>
 
 <!-- !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !! This file is generated by oca-gen-addon-readme !!
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
-<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_multiplier_free_product"><img alt="OCA/sale-promotion" src="https://img.shields.io/badge/github-OCA%2Fsale--promotion-lightgray.png?logo=github" /></a> <a class="reference external" href="https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-sale_coupon_multiplier_free_product"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external" href="https://runboat.odoo-community.org/webui/builds.html?repo=OCA/sale-promotion&amp;target_branch=15.0"><img alt="Try me on Runboat" src="https://img.shields.io/badge/runboat-Try%20me-875A7B.png" /></a></p>
-<p>This module extends from <em>coupon_multiplier_free_product</em> and allows to make
+<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/OCA/sale-promotion/tree/15.0/coupon_multiplier_free_product"><img alt="OCA/sale-promotion" src="https://img.shields.io/badge/github-OCA%2Fsale--promotion-lightgray.png?logo=github" /></a> <a class="reference external" href="https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-coupon_multiplier_free_product"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external" href="https://runboat.odoo-community.org/webui/builds.html?repo=OCA/sale-promotion&amp;target_branch=15.0"><img alt="Try me on Runboat" src="https://img.shields.io/badge/runboat-Try%20me-875A7B.png" /></a></p>
+<p>Base module that sets the configurations to create
 promotions of the type 3x2 or take this product and get
 this one another that apply as many times as the promotion rules fulfill.</p>
 <p><strong>Table of contents</strong></p>
 <div class="contents local topic" id="contents">
 <ul class="simple">
-<li><a class="reference internal" href="#usage" id="id1">Usage</a></li>
+<li><a class="reference internal" href="#configuration" id="id1">Configuration</a></li>
 <li><a class="reference internal" href="#bug-tracker" id="id2">Bug Tracker</a></li>
 <li><a class="reference internal" href="#credits" id="id3">Credits</a><ul>
 <li><a class="reference internal" href="#authors" id="id4">Authors</a></li>
@@ -384,17 +384,42 @@ this one another that apply as many times as the promotion rules fulfill.</p>
 </li>
 </ul>
 </div>
-<div class="section" id="usage">
-<h1><a class="toc-backref" href="#id1">Usage</a></h1>
-<p>Follow the configurations from the base module and be sure to guarantee the
-promotions defined conditions.</p>
+<div class="section" id="configuration">
+<h1><a class="toc-backref" href="#id1">Configuration</a></h1>
+<p>If you want to create a promotion program of the type 3x2:</p>
+<blockquote>
+<ol class="arabic simple">
+<li>Go to <em>Sales &gt; Products &gt; Promotion programs</em> and create a new one.</li>
+<li>Set the reward type to <em>Multiple of</em>.</li>
+<li>Set the <em>Free product</em> to the one in the promotion.</li>
+<li>Set the <em>Force rewarded product</em> option.</li>
+<li>Set the <em>Quantity</em> to reward.</li>
+<li>Set the <em>Quantity</em> to fulfill the reward condition.</li>
+<li>If you want a limit on how many times the product should be rewarded, set a
+<em>Maximum reward quantity</em>.</li>
+</ol>
+</blockquote>
+<p>If you want to create a promotion program of the type take one of these products and
+get this one another for free:</p>
+<blockquote>
+<ol class="arabic simple">
+<li>Go to <em>Sales &gt; Products &gt; Promotion programs</em> and create a new one.</li>
+<li>Set the reward type to <em>Multiple of</em>.</li>
+<li>Set the <em>Free product</em> to the one in the promotion.</li>
+<li>Set the <em>Quantity</em> to reward.</li>
+<li>Set the <em>Quantity</em> to fulfill the reward condition.</li>
+<li>Set the products domain to fulfill the reward condition.</li>
+<li>If you want a limit on how many times the product should be rewarded, set a
+<em>Maximum reward quantity</em>.</li>
+</ol>
+</blockquote>
 </div>
 <div class="section" id="bug-tracker">
 <h1><a class="toc-backref" href="#id2">Bug Tracker</a></h1>
 <p>Bugs are tracked on <a class="reference external" href="https://github.com/OCA/sale-promotion/issues">GitHub Issues</a>.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-<a class="reference external" href="https://github.com/OCA/sale-promotion/issues/new?body=module:%20sale_coupon_multiplier_free_product%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
+<a class="reference external" href="https://github.com/OCA/sale-promotion/issues/new?body=module:%20coupon_multiplier_free_product%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
 <p>Do not contact contributors directly about support or help with technical issues.</p>
 </div>
 <div class="section" id="credits">
@@ -424,7 +449,7 @@ mission is to support the collaborative development of Odoo features and
 promote its widespread use.</p>
 <p>Current <a class="reference external" href="https://odoo-community.org/page/maintainer-role">maintainer</a>:</p>
 <p><a class="reference external" href="https://github.com/chienandalu"><img alt="chienandalu" src="https://github.com/chienandalu.png?size=40px" /></a></p>
-<p>This module is part of the <a class="reference external" href="https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_multiplier_free_product">OCA/sale-promotion</a> project on GitHub.</p>
+<p>This module is part of the <a class="reference external" href="https://github.com/OCA/sale-promotion/tree/15.0/coupon_multiplier_free_product">OCA/sale-promotion</a> project on GitHub.</p>
 <p>You are welcome to contribute. To learn how please visit <a class="reference external" href="https://odoo-community.org/page/Contribute">https://odoo-community.org/page/Contribute</a>.</p>
 </div>
 </div>
