+++ coupon_multiplier_free_product/tests/__init__.py
@@ -1,0 +1,2 @@
+from . import test_coupon_multiplier_free_product
+from .test_coupon_multiplier_free_product import TestCouponMultiplier
