+++ sale_coupon_multiplier_free_product/models/__init__.py
@@ -1,4 +1,2 @@
-from . import sale_coupon
-from . import sale_coupon_program
-from . import sale_coupon_reward
 from . import sale_order
+from . import coupon_program
