+++ sale_coupon_multiplier_free_product/models/coupon_program.py
@@ -1,22 +1,10 @@
 # Copyright 2021 Tecnativa - David Vidal
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-from odoo import _, api, models
+from odoo import _, models
 
 
-class SaleCouponProgram(models.Model):
+class CouponProgram(models.Model):
     _inherit = "coupon.program"
-
-    @api.onchange("reward_type")
-    def _onchange_reward_type_multiple_of(self):
-        """We don't want this flag active in other options"""
-        if self.reward_type != "multiple_of":
-            self.force_rewarded_product = False
-
-    @api.onchange("reward_product_id")
-    def _onchange_reward_product_multiple_of(self):
-        """We need this to ensure some filters"""
-        if self.reward_type == "multiple_of":
-            self.discount_line_product_id = self.reward_product_id
 
     def _check_promo_code(self, order, coupon_code):
         message = super()._check_promo_code(order, coupon_code)
@@ -54,7 +42,7 @@ class SaleCouponProgram(models.Model):
             # Avoid program if 1 ordered foo on a program '1 foo, 1 free foo'
             if (
                 program.promo_applicability == "on_current_order"
-                and program._is_valid_product(program.reward_product_id)
+                and program._get_valid_products(program.reward_product_id)
                 and program.reward_type == "multiple_of"
             ):
                 ordered_rule_products_qty -= program.reward_product_quantity
@@ -73,19 +61,6 @@ class SaleCouponProgram(models.Model):
                 continue
             programs |= program
         return programs
-
-    def _is_valid_product(self, product):
-        """This method was removed, it's functionallity
-        essentially covered by _get_valid_products method"""
-        if self.force_rewarded_product:
-            return product == self.reward_product_id
-        res = super()._get_valid_products(product)
-        return product in res
-
-    def _get_valid_products(self, products):
-        if self.force_rewarded_product:
-            return products.filtered(lambda x: x == self.reward_product_id)
-        return super()._get_valid_products(products)
 
     def action_view_sales_orders(self):
         """Tracking of multiple of promotions"""
@@ -107,9 +82,7 @@ class SaleCouponProgram(models.Model):
     def _compute_order_count(self):
         """Multiple programs need a different domain to compute the order count"""
         multiple_of_programs = self.filtered(lambda x: x.reward_type == "multiple_of")
-        res = super(
-            SaleCouponProgram, self - multiple_of_programs
-        )._compute_order_count()
+        res = super(CouponProgram, self - multiple_of_programs)._compute_order_count()
         product_data = self.env["sale.order.line"].read_group(
             [
                 (
