+++ sale_coupon_multiplier_free_product/readme/USAGE.rst
@@ -1,9 +1,2 @@
-If a coupon belongs to a eCommerce restricted program, if a user tries to validate
-it from the sale order backend, it will raise an error, disallowing it.
-
-The same coupon will be usable from the eCommerce though.
-Configure the promotions as described in the previous section and then place a
-quotation that fulfills the rules.
-
-The reward will show up an extra order line, with the price unit of the applicable
-pricelist and a discount of 100%.
+Follow the configurations from the base module and be sure to guarantee the
+promotions defined conditions.
