+++ setup/coupon_multiplier_free_product/odoo/addons/coupon_multiplier_free_product
@@ -1,0 +1,1 @@
+../../../../coupon_multiplier_free_product
